@@ -4,15 +4,15 @@
     "license": "AGPL-3.0-or-later",
     "config": {
         "platform": {
-            "php": "8.1"
+            "php": "8.2"
         },
         "preferred-install": "dist",
         "optimize-autoloader": true,
         "discard-changes": true
     },
     "require": {
-        "dbp/relay-blob-library": "^0.2.4",
-        "php": ">=8.1"
+        "dbp/relay-blob-library": "^0.3.0",
+        "php": ">=8.2"
     },
     "require-dev": {
         "phpstan/phpstan": "^2.0",
